+++ tests/python_tests/conftest.py
@@ -2,12 +2,17 @@ import os
 import gc
 import pytest
 import shutil
+import subprocess  # nosec B404
 import logging
+from pathlib import Path
+
 from utils.constants import (
     get_ov_cache_dir,
     get_ov_cache_downloaded_models_dir,
     get_ov_cache_converted_models_dir,
 )
+from utils.atomic_download import AtomicDownloadManager
+from utils.network import retry_request
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -62,6 +67,47 @@ def pytest_addoption(parser):
 
 def pytest_configure(config: pytest.Config):
     pytest.selected_model_ids = config.getoption("--model_ids", default=None)
+
+
+IMAGE_GEN_MODELS = {
+    "tiny-random-latent-consistency": "echarlaix/tiny-random-latent-consistency",
+    "tiny-random-flux": "optimum-intel-internal-testing/tiny-random-flux",
+}
+
+DEFAULT_IMAGE_GEN_MODEL_ID = "tiny-random-latent-consistency"
+
+
+@pytest.fixture(scope="module")
+def image_generation_model(request):
+    model_id = getattr(request, "param", DEFAULT_IMAGE_GEN_MODEL_ID)
+    model_name = IMAGE_GEN_MODELS[model_id]
+    models_dir = get_ov_cache_converted_models_dir()
+    model_path = Path(models_dir) / model_id / model_name
+
+    manager = AtomicDownloadManager(model_path)
+
+    def convert_model(temp_path: Path) -> None:
+        command = [
+            "optimum-cli",
+            "export",
+            "openvino",
+            "--model",
+            model_name,
+            "--trust-remote-code",
+            "--weight-format",
+            "fp16",
+            str(temp_path),
+        ]
+        logger.info(f"Conversion command: {' '.join(command)}")
+        retry_request(lambda: subprocess.run(command, check=True, text=True, capture_output=True))
+
+    try:
+        manager.execute(convert_model)
+    except subprocess.CalledProcessError as error:
+        logger.exception(f"optimum-cli returned {error.returncode}. Output:\n{error.output}")
+        raise
+
+    return str(model_path)
 
 
 @pytest.fixture(scope="module", autouse=True)
